+++ src/parser/ELR/model/parsing.ts
@@ -1,4 +1,8 @@
-import type { GeneralTokenDataBinding, ILexer, Token } from "../../../lexer";
+import type {
+  GeneralTokenDataBinding,
+  ITrimmedLexer,
+  Token,
+} from "../../../lexer";
 import type { ASTNode } from "../../ast";
 import type { State } from "../DFA";
 import type { Callback, GrammarRuleContext } from "./context";
@@ -41,7 +45,7 @@ export type ParsingState<
     ErrorType,
     Token<LexerDataBindings, LexerErrorType>
   >[];
-  lexer: ILexer<LexerDataBindings, LexerActionState, LexerErrorType>;
+  lexer: ITrimmedLexer<LexerDataBindings, LexerActionState, LexerErrorType>;
 };
 
 // TODO: rename to reLexState