+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   - Feat: allow multiple calls for `Builder.state` and `Builder.error`.
   - Feat: add `ITrimmedLexer/IReadonlyTrimmedLexer`.
   - Fix: `Lexer.trimStart` shouldn't collect non-muted errors.
+- Parser
+  - **_Breaking Change_**: lexer will be trimmed after every parse. This should improve the performance.
+    - `IReadonlyTrimmedLexer.lexer` will be `IReadonlyTrimmedLexer` instead of `IReadonlyLexer`.
 
 ## v0.14.0